Updated _EVOLVE_DSIG code
Now one can set the precompilator variable _EVOLVE_DSIG to 0 or 1.

If it is 1, gsig_r, fsig_r will be evolved as independent variables. If it is 0, they will not.

No big differences in the stability of the code around r=0 is noted so far, hence none of the two versions is preferred yet (and that is why one can still choose between them; more tests are needed).