NaN detection and more work on smoothening
- bim-solver-BSSN.cpp
Added the precompilator variable _DETECT_NAN. If True, it will track the metric functions, the conformal factors, the components and the trace of the extrinsic curvatures. If False, it will track only gA.

- gridDriver.h
Fixed small bug in smoothenGF. nFrom can be negative, therefore in such a case it must be set equal to 0 in the loops over n, otherwise smoothing of primary fields does not work. Perhaps further checks are needed on this.

- Smoothening now works fairly well. Started trials to see on what fields it is better to perform it, with maximal slicing.