More on smoothening
Defined a new class SavitzkyGolayFilter in the file SavitzkyGolay header. It requires a function from Numerical Recipes, so it does not work yet. Its purpose is to compute the SG coefficients during runtime.

Many trials were done with smoothening different functions. The radial derivatives on q, if smoothened, make the integration to last only 2 seconds. They must be treated introducing up/down-wind stencils. To be done tomorrow (Mathematica is already set up for that).