Completed work on convective derivatives
- bim-solver-BSSN.cbp and bim-solver-BSSN.cpp
All the variables and files referring to the standard gauge are renamed from "BM" (standing for Bona-Masso) to "SG" (standing fo Standard Gauge).  This change is independent from the work on convective derivatives.

- eomBSSNEvolutionDR.h
The evolution equations including gDsig and fDsig are now compatible with the convective derivatives. Different trials show that the code is more stable if gDsig and fDsig are not evolved. Further investigation is needed, however.